fix: some refactor, reorg and better naming
