+++ src/models/webhook_model.ts
@@ -1,4 +1,4 @@
-import { Request } from "express";
+
 export interface Webhook {
     action: string
     receivedAt: number
