+++ src/tests/queue.test.ts
@@ -1,12 +1,11 @@
 import {
     PR_HOOK_CLOSED_1,
-    PR_HOOK_HEADER_1,
     PR_HOOK_OPENED_1,
 } from "./fixtures";
-import { Queue } from "../queue/queue";
-import { Webhook, createHook } from "../models/webhook_model";
-import { QueueManager, processQueue } from "../queue/queue_mgr";
-import { addHookToMgr } from "../services/webhooks";
+import { Queue } from "../models/queue_model";
+import { Webhook, createHook} from "../models/webhook_model";
+import { QueueManager, processQueue } from "../services/queue_mgr_service";
+import { queueHook } from "../services/webhooks_service";
 
 describe("adding to queue", () => {
     const queue: Queue<Webhook> = new Queue();
@@ -34,7 +33,7 @@ describe("test queue mgr", () => {
 
 describe("integration test of addHookToMgr", () => {
     it("should return a valid hook", async () => {
-        const hook: Webhook = await addHookToMgr(PR_HOOK_CLOSED_1);
+        const hook: Webhook = await queueHook(PR_HOOK_CLOSED_1);
         
         expect(hook.action).toEqual(PR_HOOK_CLOSED_1.action);
         expect(new Date(hook.receivedAt * 1000)).toBeInstanceOf(Date);
